pyproject.toml
[tool.poetry]
name = "llm-helpers"
version = "0.1.0"
description = "A library with functions for interacting with various LLMs"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "llm_helpers", from = "src"}]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.124.2"
python-multipart = "^0.0.20"
langchain-openai = "^1.1.2"
langchain-groq = "^1.1.0"
langchain-core = "^1.1.3"
python-dotenv = "^1.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
black = "^25.12.0"
ruff = "^0.14.9"
mypy = "^1.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
